#244 feat: Header mobile, desktop 기초 공사
